// final exit project/Angular/Frontend/src/app/models/ticketRegister.ts
export class TicketRegister {
  id: string;
  requestType: string;
  priority: string;
  travelCity: string;
  fromLocationCity: string;
  travelStartDate: Date;
  travelEndDate: Date;
  passportNumber: string;
  projectName: string;
  expenseBorneBy: string;
  travelApproverName: string;
  expectedDurationOfTravel: string;
  upperBound: string;
  anyAdditionalDetails: string;
}